New version of meshd that uses nl80211 instead of raw sockets.
This version starts the mesh on a mesh interface and passes an RSN
information element.  You can launch two instances, attach each one to a
different hwsim/mesh interface and confirm that the two establish a peer
link.  All meshd is doing at this point is start the mesh on the
interfaces, the rest is all happening in the kernel.

Also, remove files that were checked in by mistake in the first commit.