read HT info from wiphy
We start maintaining wiphy info in nlcfg. This is wrong, and should be
factored out later into some sort of meshd hwinfo/state struct. On the
bright side, scaffolding is now in place to grab the rates. We shouldn't
store those in nlcfg either...

Also, give meshd-nl80211 a vim modeline, and rename request_mesh_caps to
get_wiphy.