set wiphy channel and channel type based on config
Probably works, I'm unable to test this right now since the iface has to
be down while we change channels, and it doesn't look like this is
possible from NL80211..

Also, now freq is in nlcfg and channel_type in meshd_conf, this
indicates some refactoring needs to be done, probably into a third
structure.

Maybe the easiest thing to do is just maintain the mesh interface in
meshd-nl80211. This would lessen the burden on the user, and allow us to
bring it down or up at will. Still not sure how to do this though.