Add rudimentary support for cmake
This is just mirroring the way the linux Makefiles handled things.

libnl on Openwrt is no longer a special case, we pick up LDFLAGS through
pkg-config.

This does not support cmake on freebsd, although an extra case on the
CMAKE_SYSTEM_NAME conditional could be easily used to handle freebsd.