install_key: clean up error handling
Put the key freeing at in the error path of the function and thereby
fix a memory leak if the msg allocation succeeds but not the key
allocation.

This is still a bit non-obvious due to too many early returns, and
objects being freed in the middle of the function and by send_nlmsg(),
but perhaps its a step in the right direction.

Signed-off-by: Bob Copeland <me@bobcopeland.com>