ampe: remove Linux-specific enums
As pointed out by Franz Bettag, the ampe code uses a Linux enum
for HT operation type.  Simply rename these so that the FreeBSD
implementation can use its own settings here.  It is currently
assumed and documented that the renamed enum has the same values
as those in Linux so we can directly use them for the nl80211
API.

Signed-off-by: Bob Copeland <me@bobcopeland.com>