meshd-nl80211: add re-key support
Re-key checks that the peer-to-peer link encryption keys have been
programmed correctly on both sides of the link.

This is needed for chips/drivers that have problems with new peer-to-peer
link encryption keys being programmed. Especially ath9k is relevant
since it has exactly this problem when it's under load.

When a peer node has been authenticated (the peer-to-peer link encryption
keys have been programmed), then the re-key process is started:
- An UDP packet containing the peer's MAC address, the node's MAC address,
  IP address and UDP reply port is sent to a multicast address.
- This UDP packet is received by all neighbours and only the peer node
  with the same MAC address as the peer MAC address as in the UDP packet
  replies with an UDP packet to the IP address and UDP port combination
  as in the received UDP packet.
- The node that started the re-key process receives that UDP packet and
  conclude that the peer-to-peer link encryption keys have been installed
  correctly.

The sending of the UDP packets to the multicast address is retried a few
times when no reply is received from the correct node within a certain
timeout.

When the maximum number of retries has been exceeded then the node
concludes that the peer-to-peer link encryption keys have not been
installed correctly and therefore restarts the authentication process
for the peer.

When the node has concluded that the peer-to-peer link encryption keys
have been installed correctly, but it's still receiving UDP packets on
the multicast address from the corresponding peer node, then the other
side of the link has not installed the peer-to-peer link encryption keys
correctly and the node therefore restarts the authentication process for
the peer.

This patch was originally written by Alexis Green and then further
adjusted by me.

Originally-by: Alexis Green <agreen@cococorp.com>
Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>