meshd-nl80211: rekey: always enable responder parts
Everything of rekey is now always enabled, except the 'verify peer' part,
which is only enabled when the rekey configuration parameter is
properly set.

This allows nodes with rekey disabled and nodes with rekey enabled to
interoperate: a node that has rekey disabled will still respond to the
probes of the nodes that have rekey enabled. It just doesn't send out
probes itself.

Signed-off-by: Ferry Huberts <ferry.huberts@pelagic.nl>