fix size of auto_open_plinks attribute
Kernel was reporting on mesh up:

netlink: 'meshd-nl80211': attribute type 7 has an invalid length.

The cause is that auto_open_plinks was being set as a u32 but
is defined as a u8 in the kernel, going back to its introduction.
Fix it by using the right size variable in the nl80211 attribute.