nl80211: add stations only when rates are known
cfg80211 requires that the rates and HT/VHT caps do not change when updating
a station; however today we call NL80211_SET_STATION with HT capabilities
which results in a warning message in the debug log.

Over time there has been some confusion about what 'new_unauthenticated_peer'
should do; originally it was considered to be the first time a peer is
added to the kernel.  Since SAE and AMPE can complete without a beacon ever
being processed, there was the potential for not having a station uploaded
at all when it came time to insert the keys, which was fixed in 13ea8a8.

Unfortunately, the problem with adding a station when the first
authentication frame comes in is that these frames won't have any IEs so we
won't know the rateset at that point in time.  While we can fix the kernel
here to match the AP/STA state machine (allow changes for unauth STAs),
let's address this in userspace so it will work with existing kernels.

This partially reverts 13ea8a8 and instead uploads the sta only when a
beacon is received (as previously), or when reaching estab if we haven't
uploaded it previously.  From then on, the SET_STATION calls to set the peer
authenticated and upload the key should work.