fix IGTK check
In meshd-nl80211 we would check whether peer_igtk was set before
uploading that key into the kernel.  Unfortunately, this pointer
was always valid even if IGTK is not being used, because it's a
fixed-length array inside the candidate struct.  As a result, we
would always try to upload a key which would usually fail because
the keyid was zero.

Fix this by maintaining a boolean for whether there is actually
an IGTK and use that in the caller to pass null.