fix test to work if there are other wifi devs
On my normal dev system I have lots of real wifi devices
but I would still like to run this test over hwsim.  This
presents a couple of problems:

 - phy2 / phy3 are almost never the hwsim phys, especially
   after reloading hwsim without reloading cfg80211
 - I frequently already have a mesh0 tied to another wireless
   interface.

To fix these, ask sysfs which ones are the hwsim interfaces,
and use "smesh" instead of "mesh" to avoid conflicting with
mesh0.

Also unblock rfkill before running any tests which allows
it to succeed when devices are normally initialized with rfkill
soft blocks (as they are on my system).