fix bad format specifiers in logging
It was reported that the value printed for retransmission
counter was always zero on some platforms, but it was obviously
being updated because it would properly handle unresponsive
peers.

The root cause is that for this printf format string, and lots
of others, the type arguments didn't match the format specifiers.

Add type-checking for sae_debug, and fix up all of the issues it
found.