cleanup fsm_restart()
fsm_restart re-initializes a few fields, then expects the mgmt frame
handler to eventually call peer_ampe_init() to reset the rest.

There's no real reason to delay until that point; we can just go
ahead and reinitialize the peer as soon as we know it's going to
happen anyway.  We do need to keep the peer_ampe_init() in the mgmt
frame handler though: this is for the case where the other side
sends the first POPEN.

While in here: reformat the code so it's no longer a mix of tabs
and spaces; remove the "me" parameter which is only used in an
assert.