ampe: implement user-space mesh peering manager
Add the configuration variable "is-secure" to the config file.
If set, then instead of doing AMPE (peering + SAE auth), do only
the MPM (peering, with no encryption / protection).  This is
just like operating an unencrypted mesh network via iw, except
that all of the station management happens in user space so
there is a bit more control over when and how stations are
added.