ampe: fix a missing transition in state machine
If we get an open frame, we should transition to the OPN_RCVD
state.  Failure to do this can cause one side to reach ESTAB
(due to receiving the open and confirm frames) and the other side
getting stuck in listen since there's no transition from listen
state on a confirm frame.