clang-format all of authsae
The current coding style is a terrible mess of differently-spaced
indent levels and hard tabs.  Nowadays we have tools that do an
okay job of keeping everything aligned, and this is not a battle
worth caring about so add a .clang-format file and reformat
everything accordingly.

Style is mostly facebook style except pointers are aligned with
the variable as is common in C.  I flipped a coin on 2-vs-4
space indents and it came up as 2, so you're welcome (and I'm
sorry).