ampe: fix case where two STAs get stuck in listen
We observed that after a peer link got deleted that two stations
could no longer peer.  Once they return to the LISTEN state then
there is nothing to transition them back out to OPN_SNT.  In
the kernel, this works because a new beacon reception will trigger
one side to send a popen frame, but in userspace we do not get
any new peer candidate events as long as there is a station entry.

Fix this by adding a timer in the LISTEN state that will kick off
a popen frame if nothing happens for a little while.  After enough
unsuccessful attempts, delete the peer's station entry.