ampe: deprecate start_peer_link for ampe_open_peer_link
start_peer_link is bad naming for a function users are expected
to call, and also it takes the local mac address for no reason
at all.  Make a new function that is more sensible in both
respects.  Leave the old one for the time being to preserve API
compatibility.