Fix inconsistency with 802.11 spec for sending PLINK_CLOSE frames
In the spec (802.11-2016 tables 14-2 and 14-3) it is specified that PLINK_CLOSE should be sent (`sntCLS`) on events `OPN_RJCT` and `CNF_RJCT` if we are in the `ESTAB` state. This is inconsistent with section 14.4.10, but consistent with `wpa_supplicant`. Add the additional events to the state machine.

Additionally, PR 93 added a call to `ampe_close_link()` in `restart_fsm()` which was incorrect. Remove it.