tests: run tests inside Circle CI
This adds a new circleci config and some scripts to run our
tests on every pull request.  This turns out to be somewhat
complicated because most CI-as-a-service containers don't
have access to mac80211_hwsim kernel modules, so we end up
needing to boot our own vm inside the container and run the
tests from there.