meshd: replace nlutils.c with an iw-derived version
It has always bugged me that nlutils.c is not license-compatible
with the rest of authsae, for example it says at the top:

  cozybit Inc. makes no claim that this file can be linked against
  authsae and provides no warranties on this file.  If you decide to
  use it, do it at your own risk."

This is because these nl utility routines were taken from hostapd
at a time when it was GPL, while the SAE code is 4-clause BSD and
they are not compatible.

Luckily, there are ISC-licensed versions of many of the same routines
in iw, so use that as a source and refactor meshd-nl80211.c to use
those.  The new code is simpler and drops about 300 LOC.

While in here:

 * drop the event socket; we can do everything we need with just
   a single socket;

 * make error suppression just part of the error handler
   itself, it was really only used in one place

 * blow away send_nlmsg since it barely did anything

 * add a new global meshd_ctx that can hold global variables like
   nlcfg.  We need this because nlcfg is now known after the callbacks
   are hooked up, but the callbacks all want a pointer to nlcfg.
   We might expand use of this in the future to reduce the globals
   in meshd to a single struct