ampe: allocate frame size based on max size of IEs
It's lazy to just pick 1500 as a frame size and then do no
bounds checking at all.  Slightly better is to compute how
big we need the buffer to be up front and then assert
that we got it right, which this change does.  Better
still would be to use a bounds-checking buffer abstraction,
but that is for another day.