authsae: switch to 'chandef' multiple frequency model
Currently we support 20 and 40 MHz channels, but we would like
to add support for 80, 80+80, and 160 channels.  The latter
require a lot more flexibility in defining the operating channels.

Add some initial support for this by converting existing HT20/
HT40+/- support to the chandef model used in linux.

Add a test case for 40 MHz channels in 5G to make sure we're
covering that case.