meshd: add a setting to allow skipping beacons
Sometimes we would like to test the new peer path that
starts with management frame reception (popen/auth),
rather than the path that starts with receiving a beacon.

Add config setting "auto_open_plinks" which will
short-circuit the beacon reception path when set to false.
It is true by default.