meshd: call into ampe to set peer IEs
We were not always initializing ch_width depending on where
we came from.  It seems that the proper DRY approach here
would be to base ch_width off of the peer IEs since we will
need to save those anyway in the case of peering based on
action frames.

Add a new hook that copies the elements regardless of whether
we started from a beacon or action frame, and calculate ch_width
inside there.