Replace OPN_IGNR and CNF_IGNR with REQ_RJCT
Currently, in cases where we receive open or confirm frames with mismatched link IDs, we ignore them, using non-standard events `OPN_IGNR` and `CNF_IGNR`. This leads to a deadlock if a device `A` loses its link and tries to re-establish: device `B` thinks it's still ESTAB and ignores the new attempts to peer, so device `A` will never reconnect.

In order to avoid this deadlock, we instead will reject these frames with the event `REQ_RJCT`, which is specified in the standard (802.11-2016 sections 14.3.6.2 and 14.3.7.2) as the way to reject incompatible mesh peering frames. This is not strictly correct – we should actually create a new mesh peering instance and start its state machine (section 14.3.5). However, we don't currently currently support multiple mesh peering instances, so this is a temporary workaround.

Additionally, in the standard, the `REQ_RJCT` event is only clearly defined for the `IDLE` state (section 14.4.3, and table+figure 14.2 for MPM, table+figure 14-3 for AMPE), but in practice, it appears to make sense to process it in all states, similarly to `OPN_RJCT` and `CNF_RJCT` (after all, it too is a type of rejection of an incoming mesh peering frame).