ampe: improve mesh peering IE length tracking
The mesh peering IE can have several different lengths depending
on what kind of action frame it is, which fields are available,
and whether encryption is enabled.  We had some ugly code here
to try to figure out the final length ahead of time, but this is
just duplicating work since we also have to advance the IE pointer
anyway, and there have already been cases where the length we
calculated disagreed with the number of bytes we put in the IE.

Clean all of that up by just writing the length based on the IE
pointer after all is done.