ampe: fix potential crash if cand->conf is null
We've observed some crashes that seem to be related to
plink_frame_tx() being called with a peer that exists but
cand->conf has not yet been initialized (for example, if
the mesh station has not initiated peering yet).

Fix these by never sending close frames if cand->conf is
null.