ampe: make check_frame_protection work
check_frame_protection parses the AMPE IE and checks the AES-SIV
encrypted MIC, copying keys if the MIC is ok.  That is great, but
if the MIC fails, we should discard the frame.  Unfortunately,
the return value of check_frame_protection is always ignored, and
also the value it returns is busted in the success case.

Rename it to protection_is_valid(), have it return proper bool values,
and use the return value for security++.