ampe: fix memory leak in ampe_set_peer_ies
When receiving a peering frame, we have to make a copy
of the capabilities We free all the capabilities when we free the
peer, but if we receive more than one peering frame from the
same candidate (common) then we'll leak all of the previous
values.  Check and free before overwriting the pointer.