ampe: fix estab peer link accounting
We are incrementing and decrementing the peer link count at
various places, but the current code is insufficient: for
example, SAE code might remove a peer from the peer list if a
second instance of the SAE state machine reaches ACCEPTED (this
happens when you reboot a session, for example), and the AMPE
state machine never gets any kind of notification of that.

Instead of maintaining a count which is hard to get right, simply
compute it whenever we need it by counting up the established
STAs.  This is drier code and doesn't really hurt since there
can be at most a few dozen peers in a typical network.

Add a new test case to verify the functionality.