sae: add a wrapper for cb->fin()
There are some problems with station management in the current
code, in that delete_peer() callers sometimes expect the kernel
code to also remove the station from the kernel.  As a baby step
to rationalizing this stuff, combine fin() and delete_peer calls
so that there are fewer delete_peer() calls which are easier to
reason about.