fix station handling in cb->delete_peer() and elsewhere
Clean up a number of station handling issues and document
station lifecycle a little bit more.

This fixes an issue with the AMPE state machine that was expecting
delete_peer() to actually remove the station from the kernel
(spoiler: it did not).

When the station is removed by the kernel, we don't want to
call back into the kernel, so in that case we delete only
the local version.

When SAE state machine calls finalize, we might need to delete
the station from the kernel, or not, depending on whether it
has been uploaded; use delete_peer_by_addr for this.  Use
finalize() when we are deleting the second of two ACCEPTED peers
to make sure the station gets removed from the kernel.

Rename some functions to make their purpose clearer.